Implement play, pause, next button, see details
Unfortunately, the GUI doesn't seem to be multithreaded, so when
you click "Queue Song", the GUI is frozen until you have finished
streaming the file, so the playback controls will only work if you
are not in the queue or once your song has finished being sent. I
tried making a new thread to deal with this, but the program would
crash half the time when queueing the song (otherwise it was looking
great!). Doing a bunch of Google searches, it seems that Gtk isn't
very multithread-friendly.